+++ GRDB/Core/Configuration.swift
@@ -98,6 +98,19 @@ public struct Configuration {
     /// - Default value: false
     public var acceptsDoubleQuotedStringLiterals = false
     
+    /// When true, the `Database.suspendNotification` and
+    /// `Database.resumeNotification` suspend and resume the database. Database
+    /// suspension helps avoiding the [`0xdead10cc`
+    /// exception](https://developer.apple.com/library/archive/technotes/tn2151/_index.html).
+    ///
+    /// During suspension, all database accesses but reads in WAL mode may throw
+    /// a DatabaseError of code `SQLITE_INTERRUPT`, or `SQLITE_ABORT`. You can
+    /// check for those error codes with the
+    /// `DatabaseError.isInterruptionError` property.
+    ///
+    /// [**Experimental**](http://github.com/groue/GRDB.swift#what-are-experimental-features)
+    public var observesSuspensionNotifications = false
+    
     // MARK: - Encryption
     
     #if SQLITE_HAS_CODEC