+++ GRDB/Core/DatabaseError.swift
@@ -250,6 +250,25 @@ public struct DatabaseError: Error, CustomStringConvertible, CustomNSError {
     let arguments: StatementArguments?
 }
 
+extension DatabaseError {
+    // TODO: test
+    /// Returns true if the error has code `SQLITE_ABORT` or `SQLITE_INTERRUPT`.
+    ///
+    /// Such an error can be thrown when a database has been interrupted, or
+    /// when the database is suspended.
+    ///
+    /// See `DatabaseReader.interrupt()` and `DatabaseReader.suspend()` for
+    /// more information.
+    public var isInterruptionError: Bool {
+        switch resultCode {
+        case .SQLITE_ABORT, .SQLITE_INTERRUPT:
+            return true
+        default:
+            return false
+        }
+    }
+}
+
 // CustomStringConvertible
 extension DatabaseError {
     /// :nodoc: