+++ Tests/GRDBTests/DatabaseSuspensionTests.swift
@@ -1,11 +1,11 @@
 import XCTest
 #if GRDBCUSTOMSQLITE
-import GRDBCustomSQLite
+@testable import GRDBCustomSQLite
 #else
-import GRDB
+@testable import GRDB
 #endif
 
-class DatabaseLockPreventionTests : GRDBTestCase {
+class DatabaseSuspensionTests : GRDBTestCase {
     
     private func makeDatabaseQueue(journalMode: String) throws -> DatabaseQueue {
         let dbQueue = try makeDatabaseQueue()
@@ -16,9 +16,9 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     
     // MARK: - BEGIN TRANSACTION
     
-    func testLockPreventionPreventsNewTransactionInDeleteJournalMode() throws {
+    func testSuspensionPreventsNewTransactionInDeleteJournalMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "delete")
-        dbQueue.startPreventingLock()
+        dbQueue.suspend()
         
         do {
             try dbQueue.inDatabase { db in
@@ -27,7 +27,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "BEGIN TRANSACTION")
         }
         
@@ -38,7 +38,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "BEGIN IMMEDIATE TRANSACTION")
         }
         
@@ -49,7 +49,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "BEGIN EXCLUSIVE TRANSACTION")
         }
         
@@ -60,14 +60,14 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "SAVEPOINT test")
         }
     }
     
-    func testLockPreventionDoesNotPreventNewDeferredTransactionInWALMode() throws {
+    func testSuspensionDoesNotPreventNewDeferredTransactionInWALMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "wal")
-        dbQueue.startPreventingLock()
+        dbQueue.suspend()
         
         try dbQueue.inDatabase { db in
             try db.execute(sql: "BEGIN TRANSACTION; ROLLBACK")
@@ -75,9 +75,9 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         }
     }
     
-    func testLockPreventionPreventsNewImmediateOrExclusiveTransactionInWALMode() throws {
+    func testSuspensionPreventsNewImmediateOrExclusiveTransactionInWALMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "wal")
-        dbQueue.startPreventingLock()
+        dbQueue.suspend()
         
         do {
             try dbQueue.inDatabase { db in
@@ -86,7 +86,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "BEGIN IMMEDIATE TRANSACTION")
         }
         
@@ -97,18 +97,18 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "BEGIN EXCLUSIVE TRANSACTION")
         }
     }
     
     // MARK: - COMMIT, ROLLBACK, RELEASE SAVEPOINT, ROLLBACK TRANSACTION TO SAVEPOINT
     
-    func testLockPreventionDoesNotPreventCommit() throws {
+    func testSuspensionDoesNotPreventCommit() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "BEGIN TRANSACTION")
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 try db.execute(sql: "COMMIT")
             }
         }
@@ -116,11 +116,11 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         try test(makeDatabaseQueue(journalMode: "wal"))
     }
     
-    func testLockPreventionDoesNotPreventRollback() throws {
+    func testSuspensionDoesNotPreventRollback() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "BEGIN TRANSACTION")
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 try db.execute(sql: "ROLLBACK")
             }
         }
@@ -128,11 +128,11 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         try test(makeDatabaseQueue(journalMode: "wal"))
     }
     
-    func testLockPreventionDoesNotPreventReleaseSavePoint() throws {
+    func testSuspensionDoesNotPreventReleaseSavePoint() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "SAVEPOINT test")
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 try db.execute(sql: "RELEASE SAVEPOINT test")
             }
         }
@@ -140,11 +140,11 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         try test(makeDatabaseQueue(journalMode: "wal"))
     }
     
-    func testLockPreventionDoesNotPreventRollbackSavePoint() throws {
+    func testSuspensionDoesNotPreventRollbackSavePoint() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.inDatabase { db in
                 try db.execute(sql: "SAVEPOINT test")
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 try db.execute(sql: "ROLLBACK TRANSACTION TO SAVEPOINT test")
                 try db.execute(sql: "RELEASE SAVEPOINT test")
             }
@@ -155,27 +155,27 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     
     // MARK: - SELECT
     
-    func testLockPreventionPreventsReadInDeleteJournalMode() throws {
+    func testSuspensionPreventsReadInDeleteJournalMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "delete")
         do {
             try dbQueue.inDatabase { db in
-                try db.execute(sql: "CREATE TABLE t(a);")
-                dbQueue.startPreventingLock()
+                try db.execute(sql: "CREATE TABLE t(a)")
+                dbQueue.suspend()
                 _ = try Int.fetchOne(db, sql: "SELECT COUNT(*) FROM t")
             }
             XCTFail("Expected error")
         } catch let error as DatabaseError {
             XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-            XCTAssertEqual(error.message, "Aborted due to lock prevention")
+            XCTAssertEqual(error.message, "Database is suspended")
             XCTAssertEqual(error.sql, "SELECT COUNT(*) FROM t")
         }
     }
-
-    func testLockPreventionDoesNotPreventReadInWALMode() throws {
+    
+    func testSuspensionDoesNotPreventReadInWALMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "wal")
         try dbQueue.inDatabase { db in
-            try db.execute(sql: "CREATE TABLE t(a);")
-            dbQueue.startPreventingLock()
+            try db.execute(sql: "CREATE TABLE t(a)")
+            dbQueue.suspend()
             
             try XCTAssertEqual(Int.fetchOne(db, sql: "SELECT COUNT(*) FROM t"), 0)
             
@@ -188,34 +188,34 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     
     // MARK: - INSERT
     
-    func testLockPreventionPreventsWriteInDeleteJournalMode() throws {
+    func testSuspensionPreventsWriteInDeleteJournalMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "delete")
         try dbQueue.inDatabase { db in
-            try db.execute(sql: "CREATE TABLE t(a);")
-            dbQueue.startPreventingLock()
+            try db.execute(sql: "CREATE TABLE t(a)")
+            dbQueue.suspend()
             do {
                 try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-                XCTAssertEqual(error.message, "Aborted due to lock prevention")
+                XCTAssertEqual(error.message, "Database is suspended")
                 XCTAssertEqual(error.sql, "INSERT INTO t DEFAULT VALUES")
             }
         }
     }
     
-    func testLockPreventionPreventsWriteInWALMode() throws {
+    func testSuspensionPreventsWriteInWALMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "wal")
         try dbQueue.inDatabase { db in
-            try db.execute(sql: "CREATE TABLE t(a);")
-            dbQueue.startPreventingLock()
+            try db.execute(sql: "CREATE TABLE t(a)")
+            dbQueue.suspend()
             
             do {
                 try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-                XCTAssertEqual(error.message, "Aborted due to lock prevention")
+                XCTAssertEqual(error.message, "Database is suspended")
                 XCTAssertEqual(error.sql, "INSERT INTO t DEFAULT VALUES")
             }
             
@@ -227,7 +227,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
                 XCTFail("Expected error")
             } catch let error as DatabaseError {
                 XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-                XCTAssertEqual(error.message, "Aborted due to lock prevention")
+                XCTAssertEqual(error.message, "Database is suspended")
                 XCTAssertEqual(error.sql, "INSERT INTO t DEFAULT VALUES")
             }
         }
@@ -235,19 +235,19 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     
     // MARK: - Automatic ROLLBACK
     
-    func testLockPreventionRollbacksOnPreventedWrite() throws {
+    func testSuspensionRollbacksOnPreventedWrite() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             do {
                 try dbQueue.write { db in
-                    try db.execute(sql: "CREATE TABLE t(a);")
+                    try db.execute(sql: "CREATE TABLE t(a)")
                     XCTAssertTrue(db.isInsideTransaction)
-                    dbQueue.startPreventingLock()
+                    dbQueue.suspend()
                     do {
                         try db.execute(sql: "INSERT INTO t DEFAULT VALUES")
                         XCTFail("Expected error")
                     } catch let error as DatabaseError {
                         XCTAssertEqual(error.resultCode, .SQLITE_ABORT)
-                        XCTAssertEqual(error.message, "Aborted due to lock prevention")
+                        XCTAssertEqual(error.message, "Database is suspended")
                         XCTAssertEqual(error.sql, "INSERT INTO t DEFAULT VALUES")
                     }
                     // Aborded transaction
@@ -265,7 +265,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     
     // MARK: - Concurrent Lock Prevention
     
-    func testLockPreventionAbortsDatabaseQueueAccess() throws {
+    func testSuspensionAbortsDatabaseQueueAccess() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             let semaphore1 = DispatchSemaphore(value: 0)
             let semaphore2 = DispatchSemaphore(value: 0)
@@ -289,7 +289,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             }
             let block2 = {
                 semaphore1.wait()
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 semaphore2.signal()
             }
             let blocks = [block1, block2]
@@ -301,10 +301,10 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         try test(makeDatabaseQueue(journalMode: "wal"))
     }
     
-    func testLockPreventionDoesNotPreventFurtherReadInWALMode() throws {
+    func testSuspensionDoesNotPreventFurtherReadInWALMode() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "wal")
         try dbQueue.write { db in
-            try db.execute(sql: "CREATE TABLE t(a);")
+            try db.execute(sql: "CREATE TABLE t(a)")
         }
         
         let semaphore1 = DispatchSemaphore(value: 0)
@@ -334,7 +334,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         }
         let block2 = {
             semaphore1.wait()
-            dbQueue.startPreventingLock()
+            dbQueue.suspend()
             semaphore2.signal()
         }
         let blocks = [block1, block2]
@@ -346,7 +346,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     func testWriteTransactionAbortedDuringStatementExecution() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.write { db in
-                try db.execute(sql: "CREATE TABLE t(a);")
+                try db.execute(sql: "CREATE TABLE t(a)")
             }
             
             let semaphore1 = DispatchSemaphore(value: 0)
@@ -374,7 +374,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             }
             let block2 = {
                 semaphore1.wait()
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 semaphore2.signal()
             }
             let blocks = [block1, block2]
@@ -389,7 +389,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
     func testWriteTransactionAbortedDuringStatementExecutionPreventsFurtherDatabaseAccess() throws {
         func test(_ dbQueue: DatabaseQueue) throws {
             try dbQueue.write { db in
-                try db.execute(sql: "CREATE TABLE t(a);")
+                try db.execute(sql: "CREATE TABLE t(a)")
             }
             
             let semaphore1 = DispatchSemaphore(value: 0)
@@ -433,7 +433,7 @@ class DatabaseLockPreventionTests : GRDBTestCase {
             }
             let block2 = {
                 semaphore1.wait()
-                dbQueue.startPreventingLock()
+                dbQueue.suspend()
                 semaphore2.signal()
             }
             let blocks = [block1, block2]
@@ -445,15 +445,72 @@ class DatabaseLockPreventionTests : GRDBTestCase {
         try test(makeDatabaseQueue(journalMode: "wal"))
     }
     
-    // MARK: - stopPreventingLock
+    // MARK: - resume
     
-    func testStopPreventingLock() throws {
+    func testResume() throws {
         let dbQueue = try makeDatabaseQueue(journalMode: "delete")
         try dbQueue.inDatabase { db in
-            try db.execute(sql: "CREATE TABLE t(a);")
-            dbQueue.startPreventingLock()
-            dbQueue.stopPreventingLock()
+            try db.execute(sql: "CREATE TABLE t(a)")
+            dbQueue.suspend()
+            dbQueue.resume()
             try XCTAssertEqual(Int.fetchOne(db, sql: "SELECT COUNT(*) FROM t")!, 0)
+        }
+    }
+    
+    // MARK: - journalModeCache
+    
+    // Test for internals. Make sure the journalModeCache is not set too early,
+    // especially not before user can choose the journal mode
+    func testJournalModeCache() throws {
+        do {
+            let dbQueue = try makeDatabaseQueue()
+            try dbQueue.inDatabase { db in
+                XCTAssertNil(db.journalModeCache)
+                try db.execute(sql: "PRAGMA journal_mode=truncate")
+                try db.execute(sql: "CREATE TABLE t(a)")
+                XCTAssertNil(db.journalModeCache)
+            }
+            dbQueue.suspend()
+            dbQueue.inDatabase { db in
+                XCTAssertNil(db.journalModeCache)
+                try? db.execute(sql: "SELECT * FROM sqlite_master")
+                XCTAssertEqual(db.journalModeCache, "truncate")
+            }
+        }
+        do {
+            var configuration = Configuration()
+            configuration.prepareDatabase = { db in
+                try db.execute(sql: "PRAGMA journal_mode=truncate")
+            }
+            let dbQueue = try makeDatabaseQueue(configuration: configuration)
+            dbQueue.suspend()
+            dbQueue.inDatabase { db in
+                XCTAssertNil(db.journalModeCache)
+                try? db.execute(sql: "SELECT * FROM sqlite_master")
+                XCTAssertEqual(db.journalModeCache, "truncate")
+            }
+        }
+        do {
+            let dbPool = try makeDatabasePool()
+            try dbPool.write { db in
+                XCTAssertNil(db.journalModeCache)
+                try db.execute(sql: "CREATE TABLE t(a)")
+                XCTAssertNil(db.journalModeCache)
+            }
+            dbPool.suspend()
+            try dbPool.writeWithoutTransaction { db in
+                XCTAssertNil(db.journalModeCache)
+                try db.execute(sql: "SELECT * FROM sqlite_master")
+                XCTAssertEqual(db.journalModeCache, "wal")
+            }
+            try dbPool.write { db in
+                XCTAssertEqual(db.journalModeCache, "wal")
+            }
+            try dbPool.read { db in
+                XCTAssertNil(db.journalModeCache)
+                try db.execute(sql: "SELECT * FROM sqlite_master")
+                XCTAssertNil(db.journalModeCache)
+            }
         }
     }
 }